Moved Leaderboard cmd to separate class
